Correccion de bugs 1
- Actividad de añadir parcelas a reserva ahora se reconoce correctamente en el contexto de runtime

- Id de reserva ahora se pasa correctamente a la actividad de añadir parcelas a reserva

+++ app/src/main/java/es/unizar/eina/reservapad/database/reservas/ReservaRoomDatabase.java
@@ -23,9 +23,11 @@ public abstract class ReservaRoomDatabase extends RoomDatabase {
         if (INSTANCE == null) {
             synchronized (ReservaRoomDatabase.class) {
                 if (INSTANCE == null) {
+                    // Borra la base de datos existente en modo debug
+                    context.deleteDatabase("reserva_database");
                     INSTANCE = Room.databaseBuilder(context.getApplicationContext(),
-                                    ReservaRoomDatabase.class, "parcela_database")
-                            .fallbackToDestructiveMigration() // Permite la migración de la base de datos
+                                    ReservaRoomDatabase.class, "reserva_database")
+                            .fallbackToDestructiveMigration()
                             .addCallback(sRoomDatabaseCallback)
                             .build();
                 }
@@ -34,25 +36,27 @@ public abstract class ReservaRoomDatabase extends RoomDatabase {
         return INSTANCE;
     }
 
-    private static Callback sRoomDatabaseCallback = new Callback() {
+    private static final Callback sRoomDatabaseCallback = new Callback() {
         @Override
         public void onCreate(@NonNull SupportSQLiteDatabase db) {
             super.onCreate(db);
 
-            // If you want to keep data through app restarts,
-            // comment out the following block
             databaseWriteExecutor.execute(() -> {
-                // Populate the database in the background.
-                // If you want to start with more notes, just add them.
-                ReservaDao dao = INSTANCE.reservaDao();
-                dao.deleteAll();
+                // Obtenemos el DAO de la instancia ya construida
+                ReservaRoomDatabase database = INSTANCE; // Asegúrate de que INSTANCE ya está inicializada
+                if (database != null) {
+                    ReservaDao dao = database.reservaDao();
 
-                Reserva reserva1 = new Reserva("Juan Luis Gonzalez", 676686696,"7/11/2023", "8/11/2023");
-                dao.insert(reserva1);
-                Reserva reserva2 = new Reserva("Maria Perez", 676686697, "9/11/2023" , "10/11/2023");
-                dao.insert(reserva2);
-                Reserva reserva3 = new Reserva("Pedro Martinez", 676686698, "11/11/2023" , "12/11/2023");
-                dao.insert(reserva3);
+                    // Insertar datos iniciales
+                    dao.deleteAll();
+
+                    Reserva reserva1 = new Reserva("Juan Luis Gonzalez", 676686696,"7/11/2023", "8/11/2023");
+                    dao.insert(reserva1);
+                    Reserva reserva2 = new Reserva("Maria Perez", 676686697, "9/11/2023" , "10/11/2023");
+                    dao.insert(reserva2);
+                    Reserva reserva3 = new Reserva("Pedro Martinez", 676686698, "11/11/2023" , "12/11/2023");
+                    dao.insert(reserva3);
+                }
             });
         }
     };
